Changed movement command so that only one move needs to be issued to move forward.

+++ SoarSuite/Environments/Soar2D/src/soar2d/player/Player.java
@@ -1,7 +1,6 @@
 package soar2d.player;
 
 import java.awt.geom.Point2D;
-import java.awt.geom.Point2D.Float;
 import java.util.logging.*;
 
 import soar2d.*;
@@ -352,12 +351,12 @@ public class Player {
 		// this is for tosca's reward system
 	}
 	
-	private Point2D.Float velocity = new Point2D.Float(0,0);
-	public void setVelocity(Point2D.Float velocity) {
+	private Point2D.Double velocity = new Point2D.Double(0,0);
+	public void setVelocity(Point2D.Double velocity) {
 		assert velocity != null;
 		this.velocity = velocity;
 	}
-	public Point2D.Float getVelocity() {
+	public Point2D.Double getVelocity() {
 		return this.velocity;
 	}
 }
